+++ src/dataLayer__init/index.d.ts
@@ -1,0 +1,3 @@
+import type { gtm_script_html_opts__T } from '../gtm_script_html_opts___T'
+export declare function dataLayer__init(opts?:gtm_script_html_opts__T):void;
+export { dataLayer__init as init__dataLayer }