+++ src/init_dataLayer/index.d.ts
@@ -1,3 +1,0 @@
-import { gtm_script_html_opts__T } from '../gtm_script_html_opts___T'
-export declare function init_dataLayer(opts?:gtm_script_html_opts__T):void;
-export { init_dataLayer as init__dataLayer }