+++ src/init_dataLayer/index.test.ts
@@ -1,6 +1,0 @@
-import { test } from 'uvu'
-import { init_dataLayer } from '../index.js'
-test('init_dataLayer', ()=>{
-	init_dataLayer({ GTM_ID: 'trackme' })
-})
-test.run()